requirements.txt
flask
opencv-python
onvif-zeep
ffmpeg-python
flask-socketio